# openapi/kindred-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kindred API",
    "description": "Open, transparent dating platform API. No hidden ranking, no feature paywalls, auditable code.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/preferences": {
      "get": {
        "tags": [
          "discovery-controller"
        ],
        "operationId": "preferences",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "discovery-controller"
        ],
        "operationId": "updatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/media/profile-photo-uploads": {
      "post": {
        "tags": [
          "media-upload-controller"
        ],
        "operationId": "createProfilePhotoUpload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePhotoUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilePhotoUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/likes": {
      "post": {
        "tags": [
          "discovery-controller"
        ],
        "operationId": "react",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReactResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{id}/read": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "markRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{id}/messages": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "send",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{id}/media-uploads": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "createMediaUpload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMediaUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMediaUploadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/discovery": {
      "get": {
        "tags": [
          "discovery-controller"
        ],
        "operationId": "discover",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryCard"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/conversations/{id}/media/{mediaId}": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "mediaUrls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "mediaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMediaUrlsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdatePreferencesRequest": {
        "type": "object",
        "properties": {
          "distanceKm": {
            "type": "integer",
            "format": "int32",
            "maximum": 500,
            "minimum": 1
          },
          "ageMin": {
            "type": "integer",
            "format": "int32",
            "maximum": 120,
            "minimum": 18
          },
          "ageMax": {
            "type": "integer",
            "format": "int32",
            "maximum": 120,
            "minimum": 18
          },
          "genders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "woman",
                "man",
                "nonbinary"
              ]
            },
            "maxItems": 3,
            "minItems": 0
          },
          "lookingFor": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 0
          },
          "relationshipStyles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "monogamy",
                "non_monogamy",
                "open",
                "polyamory"
              ]
            },
            "maxItems": 4,
            "minItems": 0
          },
          "dealbreakers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 0
          },
          "weights": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "required": [
          "ageMax",
          "ageMin",
          "distanceKm"
        ]
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "ageMin": {
            "type": "integer",
            "format": "int32"
          },
          "ageMax": {
            "type": "integer",
            "format": "int32"
          },
          "genders": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "woman",
                "man",
                "nonbinary"
              ]
            }
          },
          "lookingFor": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "relationshipStyles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "monogamy",
                "non_monogamy",
                "open",
                "polyamory"
              ]
            }
          },
          "dealbreakers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "weights": {
            "$ref": "#/components/schemas/Weights"
          }
        },
        "required": [
          "ageMax",
          "ageMin",
          "dealbreakers",
          "distanceKm",
          "genders",
          "lookingFor",
          "relationshipStyles",
          "weights"
        ]
      },
      "Weights": {
        "type": "object",
        "properties": {
          "interests": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "activity": {
            "type": "number",
            "format": "double"
          },
          "mutualFit": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "activity",
          "distance",
          "interests",
          "mutualFit"
        ]
      },
...
      "PhotoUrls": {
        "type": "object",
        "properties": {
          "thumb": {
            "type": "string"
          },
          "card": {
            "type": "string"
          },
          "full": {
            "type": "string"
          }
        },
        "required": [
          "card",
          "full",
          "thumb"
        ]
      },
      "ProfilePhotoUploadRequest": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "contentType"
        ]
      },
      "ProfilePhotoUploadResponse": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "storageKey": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "expiresAt",
          "storageKey",
          "uploadUrl"
        ]
      },
      "ReactRequest": {
        "type": "object",
        "properties": {
          "toUserId": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "type": "string",
            "enum": [
              "like",
              "superlike",
              "pass"
            ]
          }
        },
        "required": [
          "kind",
          "toUserId"
        ]
      },
      "ReactResponse": {
        "type": "object",
        "properties": {
          "matched": {
            "type": "boolean"
          },
          "matchId": {
            "type": "integer",
            "format": "int64"
          },
          "conversationId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "matched"
        ]
      },
      "SendMessageRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "mediaStorageKey": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        }
      },
      "ChatMediaSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "nsfw": {
            "type": "boolean"
          },
          "blurhash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nsfw",
          "status"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "senderId": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "media": {
            "$ref": "#/components/schemas/ChatMediaSummary"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "id",
          "senderId"
        ]
      },
      "ChatMediaUploadRequest": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "contentType"
        ]
      },
      "ChatMediaUploadResponse": {
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string"
          },
          "storageKey": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "expiresAt",
          "storageKey",
          "uploadUrl"
        ]
      },
...
      "PhotoSummary": {
        "type": "object",
        "properties": {
          "urls": {
            "$ref": "#/components/schemas/PhotoUrls"
          },
          "blurhash": {
            "type": "string"
          }
        }
      },
...
      "DiscoveryCard": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "bio": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "woman",
              "man",
              "nonbinary"
            ]
          },
          "lookingFor": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "relationshipStyles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "monogamy",
                "non_monogamy",
                "open",
                "polyamory"
              ]
            }
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "photo": {
            "$ref": "#/components/schemas/PhotoSummary"
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "whyThisPerson": {
            "$ref": "#/components/schemas/Factors"
          }
        },
        "required": [
          "age",
          "displayName",
          "interests",
          "lookingFor",
          "relationshipStyles",
          "score",
          "userId",
          "whyThisPerson"
        ]
      },
      "Factors": {
        "type": "object",
        "properties": {
          "sharedInterests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "interestScore": {
            "type": "number",
            "format": "double"
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "distanceScore": {
            "type": "number",
            "format": "double"
          },
          "daysSinceActive": {
            "type": "integer",
            "format": "int64"
          },
          "activityScore": {
            "type": "number",
            "format": "double"
          },
          "mutualFitScore": {
            "type": "number",
            "format": "double"
          },
          "weights": {
            "$ref": "#/components/schemas/Weights"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "activityScore",
          "daysSinceActive",
          "distanceScore",
          "interestScore",
          "mutualFitScore",
          "sharedInterests",
          "total",
          "weights"
        ]
      },
...
      "ChatMediaUrlsResponse": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "integer",
            "format": "int64"
          },
          "urls": {
            "$ref": "#/components/schemas/PhotoUrls"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "expiresAt",
          "mediaId",
          "urls"
        ]
      }
    }
  }
}